behaviour fixes
git-svn-id: svn+ssh://c4science.ch/source/amie/trunk@1954 64fb8e13-074e-4207-b257-2d34fb7690c6

+++ viewer/triangleGlDrawer.h
@@ -75,8 +75,8 @@ protected:
 	
 	size_t currentDisplayList ;
 	int currentSet ;
-	double fracup ; 
-	double fracdown ;
+	int fracup ; 
+	int fracdown ;
 	
 public slots:
 	void setSegmentDown(int) ;
